refactor(attribute): address review feedback on the bare IP write path [INFP-551]
`_allows_prefix` becomes a module-level function rather than a staticmethod. It
cannot be an instance method while `validate_format` remains a classmethod on the
whole attribute hierarchy, and a module-level helper matches the file's existing
shape.

The IPHost default-value tests move out of `test_manager_schema.py`, which was
already too large, into their own file. That file is now untouched by this branch.

`TestValueValidationAndNormalisation` and `TestStorageAndDerivedProperties` load
their schema once per class instead of once per test, which is the dominant cost
in the suite. Neither class needs a per-test database, so the saved record they
read is built once as well.

Attribute values are read through `get_attribute()` rather than dynamic attribute
access, and the HFID test builds a typed `NodeSchema` and looks the node up by the
literal identifier rather than by the one the query returned -- feeding the
returned value back would pass even if it were wrong.

The prefix-containment test is dropped. It asserted that a Cypher `STARTS WITH`
matches a stored `binary_address`, and the exact `binary_address` is already
pinned by the value-vertex assertions, so it proved nothing the vertex-level test
does not.
